+++ js/main.js
@@ -38,7 +38,7 @@ if (plexforweb) {
     var StatsButtonParent = "NavBar-right";
     var StatsButtonContainer = "NavBarActivityButton-container";
     var plexParentBanner = "metadata-starRatings";
-    var MinPfWVersionDisp = "4.132.3";
+    var MinPfWVersionDisp = "4.133.0";
     var MinPfWVersion = (MinPfWVersionDisp).replaceAll(".", "");
 }
 else {
@@ -53,7 +53,7 @@ else {
     var StatsButtonParent = "NavBar-right";
     var StatsButtonContainer = "NavBarActivityButton-container";
     var plexParentBanner = "metadata-starRatings";
-    var MinPfWVersionDisp = "4.129.1";
+    var MinPfWVersionDisp = "4.132.2";
     var MinPfWVersion = (MinPfWVersionDisp).replaceAll(".", "");
 }
 
@@ -222,7 +222,8 @@ function runOnReady() {
 function getPlexToken() {
     if (localStorage["myPlexAccessToken"]) {
         var plex_token = localStorage["myPlexAccessToken"];
-        utils.debug("Main (getPlexToken): plex_token fetched from localStorage - " + plex_token);
+        utils.debug("Main (getPlexToken): plex_token fetched from localStorage:");
+        utils.debug(plex_token);
         return plex_token;
     }
 }